Adapts JoglNewtKey to the latest changes in the public API of NEWT

+++ ardor3d-jogl/src/main/java/com/ardor3d/input/jogl/JoglNewtKey.java
@@ -10,9 +10,8 @@
 
 package com.ardor3d.input.jogl;
 
-import com.jogamp.newt.event.KeyEvent;
 import com.ardor3d.input.Key;
-
+import com.jogamp.newt.event.KeyEvent;
 
 public enum JoglNewtKey {
 
@@ -130,12 +129,12 @@ public enum JoglNewtKey {
     private final int _newtCode;
     private final Key _key;
 
-    private JoglNewtKey(final int newtCode, final Key key) {
+    private JoglNewtKey(final short newtCode, final Key key) {
         _newtCode = newtCode;
         _key = key;
     }
 
-    public static Key findByCode(final int newtCode) {
+    public static Key findByCode(final short newtCode) {
         for (final JoglNewtKey ak : values()) {
             if (ak._newtCode == newtCode) {
                 return ak._key;
